Cast to int ranges for LUT?s due to error in python 2.7. Also, fix bug which tries to name image bands in SREF image even when SREF is not being calculated.

+++ arcsilib/arcsisensor.py
@@ -263,7 +263,7 @@ class ARCSIAbstractSensor (object):
         elevLUTFeat = collections.namedtuple('ElevLUTFeat', ['Elev', 'Coeffs'])
         lut = list()
         elevRange = (surfaceAltitudeMax - surfaceAltitudeMin) / 100
-        numElevSteps = math.ceil(elevRange) + 1
+        numElevSteps = int(math.ceil(elevRange) + 1)
         elevVal = surfaceAltitudeMin
         for i in range(numElevSteps):
             print("Building LUT Elevation ", elevVal)
@@ -279,11 +279,11 @@ class ARCSIAbstractSensor (object):
         aotLUTFeat = collections.namedtuple('AOTLUTFeat', ['AOT', 'Coeffs'])
         lut = list()
         elevRange = (surfaceAltitudeMax - surfaceAltitudeMin) / 100
-        numElevSteps = math.ceil(elevRange) + 1
+        numElevSteps = int(math.ceil(elevRange) + 1)
         elevVal = surfaceAltitudeMin
         
         aotRange = (aotMax - aotMin) / 0.05
-        numAOTSteps = math.ceil(aotRange) + 1
+        numAOTSteps = int(math.ceil(aotRange) + 1)
         aotVal = aotMin
         
         for i in range(numElevSteps):
